Provides lowercase checking of API reingest type
As part of the payload to the API call vs. the API uri itself we can make
our check against reingest-type lowercase to allow for cases where the user
has submitted an uppercase string. It also allows for cases where our docs
might be wrong and describe an upper case string.

Resolves #325